first version with serial exposure control
Last version was the simplest possible, which just ran the sensor at maximum rate
with minimum exposure time of 11 microseconds.
This version allows a serial command to be sent, in the form of
32-bit hex integer (lower case a-f) appended with 'E' to set a
32 bit, 5 MHz exposure time counter. This count is in addition to
the 11 microseconds (6 x 5MHz clocks).
The readout rate will slow to accomodate the counter,
but continues to read out the sensor at 5 MHz video rate.
A python send utility is included to easily send the exposure control messages.